Extract the paged KV cache into its own crate models opt into
