+++ crates/burn-lm-llama/src/nn/attention/mod.rs
@@ -1,8 +1,7 @@
-mod block_pool;
-mod block_store;
-mod kv_cache;
 mod mha;
 
-pub use block_pool::*;
-pub use kv_cache::*;
 pub use mha::*;
+// The paged KV machinery (block pool, block store, KeyValueCache, LanePlan, PagedKvCache) is a
+// model-side toolkit this model chose, not something it owns — re-exported here so the rest of the
+// crate keeps its familiar paths.
+pub use burn_lm_paged_kv::*;
